[bugfix] 目的関数値を計算するメソッドのバグを修正
Field#NumOfMatchesWithTargetField() の戻り値をshortに設定していたため、一致数がshortの最大値(32767)を超えている場合にオーバーフローするエラーが発生してしまっていた。

+++ StampLib/algorithm/RandomStampSolver.cs
@@ -22,7 +22,7 @@ namespace StampLib.algorithm
             Solution current_best_solution = new Solution();
 
             // 暫定解の目的関数値
-            short best_objective_value = 0;
+            int best_objective_value = 0;
 
             // 9.5秒以内で最適解を計算する
             var sw = new System.Diagnostics.Stopwatch();
